Layouts dir + component refactoring
