+++ resources/js/Components/Logo/Logo.tsx
