+++ resources/js/Components/Pagination/Pagination.tsx
