+++ resources/js/Pages/Contacts/Create.tsx
@@ -1,5 +1,5 @@
 import { Link, usePage, useForm } from '@inertiajs/react';
-import Layout from '@/Components/Layout';
+import MainLayout from '@/Layouts/MainLayout';
 import LoadingButton from '@/Components/Button/LoadingButton';
 import TextInput from '@/Components/Form/TextInput';
 import SelectInput from '@/Components/Form/SelectInput';
@@ -176,7 +176,7 @@ const Create = () => {
  * [Learn more](https://inertiajs.com/pages#persistent-layouts)
  */
 Create.layout = (page: React.ReactNode) => (
-  <Layout title="Create Contact" children={page} />
+  <MainLayout title="Create Contact" children={page} />
 );
 
 export default Create;
