+++ resources/js/Pages/Organizations/Edit.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Head } from '@inertiajs/react';
 import { Link, usePage, useForm, router } from '@inertiajs/react';
-import Layout from '@/Components/Layout';
+import MainLayout from '@/Layouts/MainLayout';
 import DeleteButton from '@/Components/Button/DeleteButton';
 import LoadingButton from '@/Components/Button/LoadingButton';
 import TextInput from '@/Components/Form/TextInput';
@@ -192,6 +192,6 @@ const Edit = () => {
  *
  * [Learn more](https://inertiajs.com/pages#persistent-layouts)
  */
-Edit.layout = (page: React.ReactNode) => <Layout children={page} />;
+Edit.layout = (page: React.ReactNode) => <MainLayout children={page} />;
 
 export default Edit;
