+++ resources/js/Pages/Organizations/Index.tsx
@@ -1,7 +1,7 @@
 import { Link, usePage } from '@inertiajs/react';
-import Layout from '@/Components/Layout';
-import SearchFilter from '@/Components/SearchFilter';
-import Pagination from '@/Components/Pagination';
+import MainLayout from '@/Layouts/MainLayout';
+import FilterBar from '@/Components/FilterBar/FilterBar';
+import Pagination from '@/Components/Pagination/Pagination';
 import { Organization, PaginatedData } from '@/types';
 import Table from '@/Components/Table/Table';
 import { Trash2 } from 'lucide-react';
@@ -20,7 +20,7 @@ function Index() {
     <div>
       <h1 className="mb-8 text-3xl font-bold">Organizations</h1>
       <div className="flex items-center justify-between mb-6">
-        <SearchFilter />
+        <FilterBar />
         <Link
           className="btn-indigo focus:outline-none"
           href={route('organizations.create')}
@@ -60,7 +60,7 @@ function Index() {
  * [Learn more](https://inertiajs.com/pages#persistent-layouts)
  */
 Index.layout = (page: React.ReactNode) => (
-  <Layout title="Organizations" children={page} />
+  <MainLayout title="Organizations" children={page} />
 );
 
 export default Index;
