+++ resources/js/Pages/Users/Index.tsx
@@ -1,7 +1,7 @@
 import { Link, usePage } from '@inertiajs/react';
-import Layout from '@/Components/Layout';
-import SearchFilter from '@/Components/SearchFilter';
-import Pagination from '@/Components/Pagination';
+import MainLayout from '@/Layouts/MainLayout';
+import FilterBar from '@/Components/FilterBar/FilterBar';
+import Pagination from '@/Components/Pagination/Pagination';
 import { PaginatedData, User } from '@/types';
 import Table from '@/Components/Table/Table';
 import { Trash2 } from 'lucide-react';
@@ -18,7 +18,7 @@ const Index = () => {
     <div>
       <h1 className="mb-8 text-3xl font-bold">Users</h1>
       <div className="flex items-center justify-between mb-6">
-        <SearchFilter />
+        <FilterBar />
         <Link
           className="btn-indigo focus:outline-none"
           href={route('users.create')}
@@ -71,7 +71,7 @@ const Index = () => {
  * [Learn more](https://inertiajs.com/pages#persistent-layouts)
  */
 Index.layout = (page: React.ReactNode) => (
-  <Layout title="Users" children={page} />
+  <MainLayout title="Users" children={page} />
 );
 
 export default Index;
